Split the std registration translation units by operator group
The six largest std operator families registered every overload from one
translation unit; each register_overload<Op, Impl>() instantiates the
overload's wiring plan, so those files peaked at 1.05-1.73 GB of compiler
memory (GCC 14, Release, precompiled headers) and could not be built side
by side on the 3-core / 7 GB hosted macOS runners.

Implements "Registration translation units" in the operators developer
guide and "Translation-Unit Budget" in the build-system guide, both added
here: a registration translation unit is budgeted at 1 GB peak compiler
memory, <family>_impl.cpp keeps register_<family>_operators() sequencing
the groups, and <family>_impl_<group>.cpp holds one coherent group of
register_overload calls behind register_<family>_<group>_overloads(),
declared beside the family function in impl/<family>_impl.h.

- arithmetic: numeric, division, temporal, container, aggregate
- collection: mapping, sequence, aggregate, set
- comparison: equality, ordering, extremum
- conversion: scalar, collection, combine (the runtime value
  conversions stay in conversion_impl.cpp)
- stream: flow, window
- temporal: components, instants
- hgraph-analytics statistics (1.25 GB, built by the Homebrew formula and
  the Docker image): container, collection, window

Each group keeps its original relative order; the registered overload set
per family is unchanged, and registration order does not affect
resolution (candidates are rank-sorted, a tie is an ambiguity error).

tools/measure_tu_cost.py replays compile_commands.json entries and reports
each unit's peak RSS, failing when a unit is over --budget unless it is
listed with --exempt. Measured on this tree (GCC 14, Linux, Release,
shared, IPO): the split registration files sit between 0.70 and 0.92 GB
(previously arithmetic 1.73, conversion 1.40, collection 1.36, temporal
1.13, data_frame 1.09, comparison 1.08, stream 1.05 GB).
data_frame_impl.cpp (1.09 GB, Arrow header weight rather than
registrations) is the documented exception.

The Catch2 suites remain heavier (test_std_operators.cpp 2.3 GB, nine
more between 1.1 and 1.6 GB); they are outside the registration budget
and the guide records their figures. With the library units under
budget, the native-cpp and language workflows build with four jobs on
Linux and two (three for the language job, which builds no core suite)
on the 7 GB macOS runners instead of two and one.
