+++ .github/workflows/language.yml
@@ -35,9 +35,9 @@ jobs:
       matrix:
         include:
           - os: ubuntu-24.04
-            parallel: "2"
+            parallel: "4"
           - os: macos-26
-            parallel: "1"
+            parallel: "3"
 
     steps:
       - uses: actions/checkout@d23441a48e516b6c34aea4fa41551a30e30af803 # v6
