+++ extensions/analytics/tools/audit_distribution.py
@@ -34,6 +34,9 @@ SDIST_REQUIRED = (
     "src/shaped_array_operators.cpp",
     "src/shaped_array_operators.h",
     "src/statistics.cpp",
+    "src/statistics_collection.cpp",
+    "src/statistics_container.cpp",
+    "src/statistics_window.cpp",
     "test_package/CMakeLists.txt",
     "tests/test_array_operators.cpp",
     "tests/test_pct_change.cpp",
